Update layout to include a Sidebar and enhance search functionality
Replaced MenuBar with SideBar in the Layout component. Extended the search functionality to allow users to search for tweets and other users based on username. Also made UI enhancements, which includes the addition of Search page for displaying search results and updating navigation and routes to provide easy access to the new feature. The Redux State Display was also uncommented to reveal user, menu, and loading state, while lodash was added to the devDependencies.